backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
transformers==4.35.2
torch==2.1.1+cpu
torchvision==0.16.1+cpu
torchaudio==2.1.1+cpu
python-docx==1.2.0 